Módulo 6 - Banco de Dados/letscode_app/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2==2.9.3
python-dotenv==0.19.2
Werkzeug==2.0.2